fix(panel-admin): resolve owner record in check-in code actions
Extract the generate/revoke check-in code actions into dedicated classes and
fetch the event via $livewire->getOwnerRecord() — a relation-manager header
action has no $record/$model to inject, which caused 'Attempt to read property
id on null'. Also adds array-shape typing, filled() for max_uses, final classes
and consistent action names.
